RaCoTo-Backend/ChainlinkFunctionSources/PrototypesPython/pyproject.toml
[tool.poetry]
name = "proof-of-preservation"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.29.0"
openapi-python-client = "^0.13.4"
ipykernel = "^6.22.0"
notebook = "^6.5.4"
pandas = "^2.0.1"
openpyxl = "^3.1.2"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
